[EA3-199] refactor : 도메인명 수정 및 그룹채팅 로직 리팩터링
