--- src/main/java/grep/neogulcoder/domain/calender/exception/CalendarNotFoundException.java
@@ -1,11 +1,0 @@
-package grep.neogulcoder.domain.calender.exception;
-
-import grep.neogulcoder.global.exception.business.NotFoundException;
-import grep.neogulcoder.global.response.code.ErrorCode;
-
-public class CalendarNotFoundException extends NotFoundException {
-
-    public CalendarNotFoundException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
