--- src/main/java/grep/neogulcoder/domain/calender/exception/CalendarValidationException.java
@@ -1,11 +1,0 @@
-package grep.neogulcoder.domain.calender.exception;
-
-import grep.neogulcoder.domain.calender.exception.code.CalendarErrorCode;
-import grep.neogulcoder.global.exception.validation.ValidationException;
-
-public class CalendarValidationException extends ValidationException {
-
-    public CalendarValidationException(CalendarErrorCode errorCode) {
-        super(errorCode);
-    }
-}
