+++ src/main/java/grep/neogulcoder/domain/study/service/StudyService.java
@@ -1,7 +1,7 @@
 package grep.neogulcoder.domain.study.service;
 
-import grep.neogulcoder.domain.calender.controller.dto.response.TeamCalendarResponse;
-import grep.neogulcoder.domain.calender.service.TeamCalendarService;
+import grep.neogulcoder.domain.calendar.controller.dto.response.TeamCalendarResponse;
+import grep.neogulcoder.domain.calendar.service.TeamCalendarService;
 import grep.neogulcoder.domain.groupchat.entity.GroupChatRoom;
 import grep.neogulcoder.domain.groupchat.repository.GroupChatRoomRepository;
 import grep.neogulcoder.domain.recruitment.post.repository.RecruitmentPostRepository;
